project structure update
