+++ Backend_WhatToWeEat/main/__init__.py
@@ -1,0 +1,16 @@
+from flask import Flask
+import os
+from flask_sqlalchemy import SQLAlchemy
+
+# create the extension
+db = SQLAlchemy()
+# create the app
+app = Flask(__name__)
+# configure the SQLite database, relative to the app instance folder
+app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///project.db"
+# initialize the app with the extension
+db.init_app(app)
+
+
+from main import model
+from main import route
